util: adding StreamReader and Varint utils
This change adds two classes that can be used for parsing transactions and blocks:
StreamReader and Varint.

StreamReader provides a blocking interface for the read method of a stream. This
class is necessary because the standard read() method on a Readable stream will
return null if there is not information in the read buffer. You then need to wait
for the "readable" event to fire before attempting read information again. This
may not be the best way to handle this, but it will work for now.

The Varint util uses the ability to read from the stream in chunks to decode varint
values. An encode function has also been added.