script: add p2sh method to Script.evaluate
When a data block is pushed ontot the stack, the stack is evaluated to see
if the remaining commands match the p2sh pattern. If they do, the p2sh pattern
commands are evaluated. If the p2sh pattern evaluates correctly, the redeemScript
is parsed and pushed onto the commands