network: modifying SimpleNode to read partial data from the stream
This change modifies the socket to be in paused mode. When reading data from
the stream, it is possible that the header information and the payload arrive
seperately. In this event, the header is stored until additional data is
available on the socket.

Ideally this would use the unshift method to push the header bytes back onto
the stream, but I was unable to prevent the readable event from immediately
firing resulting in a tight loop.