core: adds validateMerkleRoot to block
This code checks a list of transaction hashes that are added as a property
to the block class. The merkle check requires converting between internal
byte order and RPC byte order.

This also made changes to the comments to make it more clear on internal
byte order vs RPC byte order instead of using little-endian and
big-endian, which can be more difficult to understand.