core: modify MerkleTree construction and add isValid to MerkleBlock
This code modifies the MerkleTree construction to leave out duplicate nodes.
The tree calculation looks to see if there is no right now and uses the left
node if it doesn't exist.

On the proof tree construction, if there are no hashes left and we are still
attempting to trarverse, we short circuit node creation and return undefined.
This allows the calculation (which occurs after both left and right nodes
have been processed) to pick up the left value.

Tree construction from hashes also creates a tree where right most nodes
are not duplicated. This allows for construction of a merkle proof in the
future.