script: adding evaluation for P2WSH
This evaluation checks the stack to see if it matches the P2WSH pattern
and if it does, it pulls the redeemScript from the last element of the
witness data.

It validates the sha256 of the witness data.

If it is valid, then it loads the commands in the script into the command
stack and things can go on their merry way.